check full changelog, fixed cascade delete, multiple issues fixed...
	-fixed the cascade delete for Employee, Product and Purchase. Now there
is implemented a checkIfCanDelete() for every one of said Classes.
Category doesn't have, my reasoning is that we can delete product or
category and it would not be as impactful as deleting a purchase. Also
we need to make UnitTests for these i will work on them on Sunday.
	-Added EntityAlreadyReferencedInDatabseException for said deletes
	-fixed issue with editing yourself as employee now alert is
confirmation and if user chooses not to save will not proceed, before
user could press X and it would still save and log him out.
	-Finished PurchaseTab now we have calculated column totalPrice.
	-Also added Double totalPrice to Purchase.java it is automatically
calculated based on productsInPurchase.
	-some minor issues fixed